+++ tsutils/__init__.py
@@ -1,11 +1,10 @@
-from .cogsettings import *
-from .discord_py_utils import *
+from .cog_settings import *
 from .emoji_utils import *
 from .format_utils import *
 from .helper_classes import *
 from .helper_functions import *
 from .json_utils import *
 from .menu import *
-from .misc_stuff import *
-from .pdx_utils import *
+from .pad_utils import *
 from .tsubaki_utils import *
+from .user_interaction import *
