--- tsutils/misc_stuff.py
@@ -1,8 +1,0 @@
-import pytz
-
-# TZ used for PAD NA
-# NA_TZ_OBJ = pytz.timezone('America/Los_Angeles')
-NA_TZ_OBJ = pytz.timezone('US/Pacific')
-
-# TZ used for PAD JP
-JP_TZ_OBJ = pytz.timezone('Asia/Tokyo')
